3897 - Admin manage collaborators refactor - Clean up code (#3928)
* 3897 - Remove Users List admin route logic

* 3897 - Remove UserListFilters

* 3897 - Fix ts issue in userManagementSlice

* 3897 - Remove count state & useRoleNames

* 3897 - Remove old UserManagement styles

* 3897 - Fix import & defaultPros
