--- src/client/components/UserList/AdministrationListElement/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AdministrationListElement'
