+++ src/client/components/UserList/UserListHeader/UserListHeader.tsx
@@ -1,21 +1,13 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { RoleName, Users } from 'meta/user'
-
-import { useFilteredRoleNames } from 'client/store/ui/userManagement'
-import { useIsAdminRoute } from 'client/hooks'
 import InviteUserLink from 'client/components/UserList/InviteUserLink'
-
-import UserListButtonExport from '../UserListButtonExport'
+import UserListButtonExport from 'client/components/UserList/UserListButtonExport'
 
 type Props = { readOnly: boolean }
 
 const UserListHeader: React.FC<Props> = (props) => {
   const { readOnly } = props
-  const isAdminRoute = useIsAdminRoute()
-
-  const filteredRoleNames = useFilteredRoleNames()
 
   const { t } = useTranslation()
 
@@ -23,23 +15,13 @@ const UserListHeader: React.FC<Props> = (props) => {
     <thead>
       <tr>
         <th className="user-list__header-cell">{t('common.name')}</th>
-        {!isAdminRoute && (
-          <>
-            <th className="user-list__header-cell">{t('common.role')}</th>
-            <th className="user-list__header-cell">{t('common.email')}</th>
-          </>
-        )}
-        {isAdminRoute &&
-          filteredRoleNames.map((roleName: RoleName) => (
-            <th key={roleName} className="user-list__header-cell">
-              {t(Users.getI18nRoleLabelKey(roleName))}
-            </th>
-          ))}
+        <th className="user-list__header-cell">{t('common.role')}</th>
+        <th className="user-list__header-cell">{t('common.email')}</th>
         {!readOnly && (
           <th className="user-list__header-cell user-list__edit-column">
             <div className="user-list__edit-column-buttons-container">
-              <UserListButtonExport isAdmin={isAdminRoute} />
-              {!isAdminRoute && <InviteUserLink />}
+              <UserListButtonExport />
+              <InviteUserLink />
             </div>
           </th>
         )}
