--- src/client/components/UserListFilters/UserListFilters.scss
@@ -1,45 +1,0 @@
-@import 'src/client/style/partials';
-
-.users__table-filter {
-  display: grid;
-  grid-template-columns: 100%;
-  gap: $spacing-xs;
-  border-radius: 3px;
-  margin: $spacing-xxs 20% $spacing-s;
-  padding: $spacing-s;
-  background-color: $ui-bg;
-}
-
-.users__table-filter-title {
-  justify-self: center;
-}
-
-.users__table-filter-container {
-  display: grid;
-  gap: $spacing-xs;
-}
-
-.users__table-filter-item {
-  display: grid;
-  grid-template-columns: 20% auto;
-  grid-column-gap: $spacing-xxs;
-  align-items: center;
-
-  input {
-    width: 100%;
-    border: none;
-    padding: 10px 24px 10px 10px;
-    background: white;
-    outline: $ui-thin-border;
-  }
-
-  .multi-select {
-    background: white;
-    outline: $ui-thin-border;
-  }
-}
-
-.users__table-filter-item-label {
-  align-self: center;
-  justify-self: right;
-}
