--- src/client/components/UserListFilters/index.ts
@@ -1,1 +1,0 @@
-export { default } from './UserListFilters'
