--- src/client/components/UsersCount/UsersCount.scss
@@ -1,14 +1,0 @@
-@import 'src/client/style/partials';
-
-.user-counts__container {
-  border-top: 1px solid $text-heading;
-  padding: $font-xs;
-  margin-top: -1px;
-}
-
-.user-counts__item {
-  color: $text-mute;
-  padding: 1px 0;
-  font-size: $font-xs;
-  font-weight: bold;
-}
