--- src/client/pages/Admin/UserManagement/UserManagement.scss
@@ -1,52 +1,0 @@
-@import 'src/client/style/partials';
-
-.user-paginator {
-  display: flex;
-  list-style: none;
-  justify-content: center;
-  margin-bottom: $font-l;
-
-  li {
-    border: 1px solid rgba(0, 0, 0, 0);
-    margin-left: -1px;
-    border-color: #dadada;
-
-    &.disabled {
-      &:hover {
-        background-color: darken($ui-bg, 4%);
-        color: inherit;
-      }
-
-      a {
-        cursor: default;
-      }
-    }
-
-    &:first-child {
-      margin-left: 0;
-      border-top-left-radius: 0.25rem;
-      border-bottom-left-radius: 0.25rem;
-    }
-
-    &:last-child {
-      border-top-right-radius: 0.25rem;
-      border-bottom-right-radius: 0.25rem;
-    }
-
-    &:hover {
-      background-color: darken($ui-bg, 4%);
-      color: inherit;
-    }
-
-    &.selected {
-      background-color: $ui-accent-light;
-      color: inherit;
-    }
-
-    a {
-      cursor: pointer;
-      display: block;
-      padding: 0.5rem 0.75rem;
-    }
-  }
-}
