+++ src/client/pages/Admin/UserManagement/UsersCount/UsersCount.tsx
@@ -6,13 +6,20 @@ import { ApiEndPoint } from 'meta/api/endpoint'
 import { RoleName, Users } from 'meta/user'
 
 import { useTablePaginatedCount } from 'client/store/ui/tablePaginated'
-import { useRoleNames } from 'client/store/ui/userManagement'
 import { TablePaginatedCounterComponent } from 'client/components/TablePaginated'
 
 const UsersCount: TablePaginatedCounterComponent = () => {
   const { t } = useTranslation()
 
-  const roleNames = useRoleNames()
+  const roleNames = [
+    RoleName.ADMINISTRATOR,
+    RoleName.REVIEWER,
+    RoleName.NATIONAL_CORRESPONDENT,
+    RoleName.ALTERNATE_NATIONAL_CORRESPONDENT,
+    RoleName.COLLABORATOR,
+    RoleName.VIEWER,
+  ]
+
   const usersCount = useTablePaginatedCount(ApiEndPoint.Admin.users()) as Record<RoleName | 'total', number>
 
   return (
