+++ src/client/store/ui/userManagement/actions/index.ts
@@ -1,5 +1,4 @@
 export { getUsers } from './getUsers'
-export { getUsersCount } from './getUsersCount'
 export { getUserToEdit } from './getUserToEdit'
 export { inviteUser } from './inviteUser'
 export { removeInvitation } from './removeInvitation'
