+++ src/client/store/ui/userManagement/index.ts
@@ -1,3 +1,3 @@
-export { useFilteredRoleNames, useFilters, useRoleNames, useUsers, useUsersCount } from './hooks'
+export { useUsers } from './hooks'
 export { UserManagementActions } from './slice'
 export type { UserManagementState } from './stateType'
